Refactored DSL inclusion such that the patching of RSchema::DSL is opt-in via require 'rschema_hamster/extend_rschema_dsl'
