+++ spec/environment.rb
@@ -5,6 +5,7 @@ PROJ_ROOT = File.expand_path(File.dirname(__FILE__) + "/..")
 $: << "#{PROJ_ROOT}/lib"
 
 require 'rschema_hamster'
+# require 'rschema_hamster/extend_rschema_dsl'
 
 # For testing
 require 'pry'
